Nest backend

+++ apps/api/src/app/features/manage-users/use-cases/delete.user.use.case.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { UseCase } from '../../../core/abstracts/use.case';
 import { FindOneUserParam } from '../../../endpoints/manage-users/params/find.one.user.param';
+import { InvalidInputError } from '../errors/invalid.input.error';
 import { UserNotFoundError } from '../errors/user.not.found.error';
 import { ManageUsersRepository } from '../repositories/manage.users.repository';
 
@@ -8,7 +9,11 @@ import { ManageUsersRepository } from '../repositories/manage.users.repository';
 export class DeleteUserUseCase implements UseCase<FindOneUserParam, void> {
   constructor(private readonly repository: ManageUsersRepository) {}
 
-  execute(param: FindOneUserParam): Promise<void | UserNotFoundError | Error> {
+  async execute(
+    param: FindOneUserParam
+  ): Promise<void | UserNotFoundError | InvalidInputError | Error> {
+    if (!param) return new InvalidInputError();
+
     return this.repository.deleteUser(param);
   }
 }
